requirements.txt
certifi==2023.7.22
charset-normalizer==2.0.12
idna==3.4
numpy==1.19.5
pandas==1.1.5
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3.post1
requests==2.27.1
six==1.16.0
urllib3==1.26.16
PyYAML==6.0.1
